refactor: move app class from main.app to app/app.dart file
